Returned to segmentation after prediction task
Returned to the plots and looked over the data another time
to see if I had missed anything critical.
The data from train to test looks consistent and the model I used seemed to make
adequate predictions.
A deeper analysis and more time would be required
to compare the models on a graph.
This would help me gain some insight into the next
steps.

For now I will clean up and commit the README.